+++ src/Sections/HeroPage.js
@@ -3,9 +3,9 @@ import styled from 'styled-components';
 import { Slogan } from 'components/Slogan';
 import { HeroTextContent } from 'components/HeroText';
 import Navbar from 'components/Navbar';
-import heroBg from '../assets/hero.svg';
-import heroBgLarge from '../assets/hero-large.svg';
-import logoHero from '../assets/logo-hero.svg';
+import heroBg from '../assets/background/hero.svg';
+import heroBgLarge from '../assets/background/hero-large.svg';
+import logoHero from '../assets/logo/logo-hero.svg';
 import { ButtonHover } from '../components/ButtonHover';
 import joinBtn from '../assets/button-join.svg';
 import joinBtnHover from '../assets/button-join-hover.svg';
